+++ apps/storybook/src/stories/ui/sortable/sortable-kanban-board.tsx
@@ -1,4 +1,4 @@
-import { useRef, useState } from "react";
+import { useRef, useState, type ComponentProps } from "react";
 import { CollisionPriority } from "@dnd-kit/abstract";
 import { move } from "@dnd-kit/helpers";
 import {
@@ -7,31 +7,10 @@ import {
   type DragOverEvent,
   type DragStartEvent,
 } from "@dnd-kit/react";
-import type { Meta, StoryObj } from "storybook-react-rsbuild";
 
 import { cn } from "@notion-kit/cn";
 import { Icon } from "@notion-kit/icons";
-import {
-  Badge,
-  Button,
-  MenuGroup,
-  MenuItem,
-  Sortable,
-} from "@notion-kit/ui/primitives";
-
-const meta = {
-  title: "Notion UI/Sortable",
-  parameters: { layout: "centered" },
-} satisfies Meta;
-
-export default meta;
-type Story = StoryObj<typeof meta>;
-
-const menuItems = [
-  { id: "inbox", label: "Inbox", icon: <Icon.Globe /> },
-  { id: "roadmap", label: "Roadmap", icon: <Icon.ViewBoard /> },
-  { id: "archive", label: "Archive", icon: <Icon.Clock /> },
-];
+import { Badge, Button, Sortable } from "@notion-kit/ui/primitives";
 
 interface BoardColumn {
   id: string;
@@ -148,7 +127,7 @@ function moveBoardCards(
   return haveSameCardSet(items, next) ? next : items;
 }
 
-interface BoardCardPreviewProps extends React.ComponentProps<"div"> {
+interface BoardCardPreviewProps extends ComponentProps<"div"> {
   card: BoardCard;
   overlay?: boolean;
 }
@@ -283,7 +262,7 @@ function KanbanColumn({
   );
 }
 
-function SortableKanbanBoard() {
+export function SortableKanbanBoard() {
   const [columnOrder, setColumnOrder] = useState(
     boardColumns.map((column) => column.id),
   );
@@ -366,64 +345,3 @@ function SortableKanbanBoard() {
     </div>
   );
 }
-
-function SortableMenuList({ itemHandle }: { itemHandle?: boolean }) {
-  const [items, setItems] = useState(menuItems);
-
-  return (
-    <Sortable.Root
-      items={items.map((item) => item.id)}
-      onItemsChange={(orderedIds) => {
-        const itemsById = new Map(items.map((item) => [item.id, item]));
-        setItems(
-          orderedIds.flatMap((id) => {
-            const item = itemsById.get(String(id));
-            return item ? [item] : [];
-          }),
-        );
-      }}
-    >
-      <Sortable.List
-        className="w-64 rounded-lg bg-popover p-1 shadow-sm"
-        render={<MenuGroup />}
-      >
-        {items.map((item, index) => (
-          <Sortable.Item
-            key={item.id}
-            id={item.id}
-            index={index}
-            render={
-              <MenuItem
-                icon={
-                  itemHandle ? (
-                    item.icon
-                  ) : (
-                    <div>
-                      <div className="size-5 group-hover/item:hidden">
-                        {item.icon}
-                      </div>
-                      <Sortable.Handle className="hidden size-5 group-hover/item:flex" />
-                    </div>
-                  )
-                }
-                label={item.label}
-              />
-            }
-          />
-        ))}
-      </Sortable.List>
-    </Sortable.Root>
-  );
-}
-
-export const SortableList: Story = {
-  render: () => <SortableMenuList />,
-};
-
-export const SortableSidebarList: Story = {
-  render: () => <SortableMenuList itemHandle />,
-};
-
-export const KanbanBoard: Story = {
-  render: () => <SortableKanbanBoard />,
-};
